v9.47 Wave 1: the schema and vocabulary window, spent once
Five declared batches, committed together because they touch the same files and
the database is rebuilt for them once rather than five times.

- Cancelled is removed. CardStage.CANCELLED, cancelled_at, the all-Cancelled
  derivation branch, the cancel mode, the sprint's cancelled metric, and the
  vocabulary only Cancelled read: finish_action's terminal stage, the gated flag
  on require_check_answers, record_sprint_result's stage argument.
- The middle Card kind is Subgoal, and it is always under a Goal. The stored
  word, every prompt line and every scenario. A root-level Subgoal is refused on
  creation and when its parent is taken away, so no screen offers Subgoal as a
  kind: no screen sets a parent.
- Idea is raw capture: a title and a note, outside the tree and outside every
  stage, with its own list, its own screen and one Expand button that hands it to
  Safwa as an ordinary owner turn. ai_cards is the tree; ai_ideas is the capture.
- Effort says what an Action costs the owner, not how long it takes. EFFORT_RUNGS
  carries that wording for the selector and the tool's field description alike,
  0.5 is a rung, and effort_points, capacity_effort_points and effort_snapshot
  are numeric.
- A Sprint number says when it ran: yy.MM-xx, the month it started in and its
  place in that month, following the highest ever used there.

Both snapshots are declared: the prompt prefix moved with the prompts, and the
schema moved in cards, sprints, sprint_commitments and user_profile. There are
no migrations, so the database is rebuilt.

+++ src/safwa/features/profile/api.py
@@ -26,7 +26,7 @@ async def sprint_length_days(session: AsyncSession) -> int:
     return profile.sprint_length_days if profile is not None else SPRINT_LENGTH_DAYS
 
 
-async def capacity_effort_points(session: AsyncSession) -> int | None:
+async def capacity_effort_points(session: AsyncSession) -> float | None:
     """The effort the owner means to take on in a Sprint, or None when it is off."""
     profile = await session.get(UserProfile, 1)
     return profile.capacity_effort_points if profile is not None else None
